+++ steelc/src/codegen/computation/compile_time_computator_visitor.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "../../parser/ast/ast_visitor.h"
-
-class compile_time_computator_visitor : ast_visitor {
-
-};