+++ steelc/src/error/error_catalog.h
@@ -139,6 +139,7 @@ enum error_code {
     ERR_ENUM_OPTION_MEMBER_ACCESS,
     ERR_ARRAY_SIZE_MUST_BE_INTEGER,
     ERR_ARRAY_SIZE_MUST_BE_CONSTANT,
+    ERR_SCOPED_FUNCTION_NOT_MODULE,
 };
 
 enum warning_code {
@@ -289,7 +290,8 @@ struct error_catalog {
             {"S142", "The enum '%s' already contains an option named '%s'"},
             {"S143", "Member access not allowed on enum option '%s'"},
             {"S144", "Array size must be an integer"},
-            {"S144", "Array size must be a constant compile-time computable value"},
+            {"S145", "Array size must be a constant compile-time computable value"},
+            {"S146", "Scoped function calls (e.g. scope::func()) can only be used on modules"},
         };
         return errors[code - 1 /* -1 to avoid ERR_SUCCESS */];
     }