+++ steelc/src/parser/ast/expressions/member_expression.h
@@ -5,14 +5,14 @@
 
 #include "../ast_node.h"
 #include "expression.h"
-#include "identifier_expression.h"
+#include "../../../lexer/token_type.h"
 
 class member_expression : public expression, public std::enable_shared_from_this<member_expression> {
 public:
 	ENABLE_ACCEPT(member_expression)
 
-	member_expression(std::shared_ptr<expression> object, std::string member)
-		: object(object), member(member), resolved_type(nullptr) {
+	member_expression(std::shared_ptr<expression> object, std::string member, token_type access_operator)
+		: object(object), member(member), access_operator(access_operator), resolved_type(nullptr) {
 	}
 
 	std::string string(int indent) const override {
@@ -33,7 +33,8 @@ public:
 	ast_ptr clone() const override {
 		auto cloned = std::make_shared<member_expression>(
 			std::dynamic_pointer_cast<expression>(object->clone()),
-			member
+			member,
+			access_operator
 		);
 		cloned->resolved_type = resolved_type;
 		return cloned;
@@ -51,6 +52,7 @@ public:
 
 	std::shared_ptr<expression> object;
 	std::string member;
+	token_type access_operator;
 	type_ptr resolved_type;
 	bool is_lvalue = true;
 };