+++ steelc/src/parser/modules/module_info.h
@@ -2,7 +2,6 @@
 
 #include <string>
 #include <memory>
-#include <filesystem>
 
 #include "../symbolics/symbol_table.h"