+++ steelc/src/utils/path_utils.h
@@ -5,4 +5,5 @@
 
 namespace path_utils {
 	std::filesystem::path normalize_path(const std::string& path);
+	std::filesystem::path normalize_path(const std::filesystem::path& path);
 }